requirements-worker.txt
vastai-sdk @ git+https://github.com/vast-ai/vast-sdk.git
requests==2.32.3
